Start listener protocol after meme tables loader supervisor
